Prepare for splitting io source states
