+++ src/sys/unix/mod.rs
@@ -15,7 +15,7 @@ macro_rules! syscall {
 
 cfg_os_poll! {
     mod selector;
-    pub(crate) use self::selector::{event, Event, Events, Selector};
+    pub(crate) use self::selector::{event, Event, Events, Selector, IoSourceState};
 
     mod sourcefd;
     pub use self::sourcefd::SourceFd;
@@ -30,28 +30,6 @@ cfg_os_poll! {
         pub(crate) mod udp;
         pub(crate) mod uds;
         pub use self::uds::SocketAddr;
-    }
-
-    cfg_io_source! {
-        use std::io;
-
-        // Both `kqueue` and `epoll` don't need to hold any user space state.
-        pub(crate) struct IoSourceState;
-
-        impl IoSourceState {
-            pub fn new() -> IoSourceState {
-                IoSourceState
-            }
-
-            pub fn do_io<T, F, R>(&self, f: F, io: &T) -> io::Result<R>
-            where
-                F: FnOnce(&T) -> io::Result<R>,
-            {
-                // We don't hold state, so we can just call the function and
-                // return.
-                f(io)
-            }
-        }
     }
 
     cfg_os_ext! {
